+++ ontimize/components/table/extensions/header/table-menu/o-table-menu.component.ts
@@ -355,14 +355,12 @@ export class OTableMenuComponent implements OnInit, AfterViewInit, OnDestroy {
       const self = this;
       this.dialogService.confirm('CONFIRM', 'MESSAGES.CONFIRM_DISCARD_FILTER_BY_COLUMN').then(res => {
         if (res) {
-          self.table.dataSource.clearColumnFilters();
+          self.table.clearColumnFilters();
         }
         self.table.showFilterByColumnIcon = !res;
-        self.table.cd.detectChanges();
       });
     } else {
       this.table.showFilterByColumnIcon = !this.table.showFilterByColumnIcon;
-      this.table.cd.detectChanges();
     }
   }