+++ ontimize/components/table/extensions/o-table-storage.class.ts
@@ -30,7 +30,7 @@ export class OTableStorage {
       'filter': this.table.oTableQuickFilterComponent ? this.table.oTableQuickFilterComponent.value : ''
     };
 
-    const properties = ['sort', 'columns-display', 'columns-filter', 'quick-filter', 'page', 'selection', 'initial-configuration'];
+    const properties = ['sort', 'columns-display', 'columns-filter', 'quick-filter', 'page', 'selection', 'initial-configuration', 'filter-column-active-by-default'];
 
     Object.assign(dataToStore, this.getTablePropertiesToStore(properties));
 
@@ -78,9 +78,13 @@ export class OTableStorage {
       case 'initial-configuration':
         result = this.getInitialConfigurationState();
         break;
+      case 'filter-column-active-by-default':
+        result = this.getFilterColumnActiveByDefaultState();
+        break;
     }
     return result;
   }
+
 
   reset() {
     let state = {};
@@ -102,6 +106,14 @@ export class OTableStorage {
         sortColumns.push(sortData.id + Codes.COLUMNS_ALIAS_SEPARATOR + sortData.direction);
       });
       result['sort-columns'] = sortColumns.join(Codes.ARRAY_INPUT_SEPARATOR);
+    }
+    return result;
+  }
+
+  protected getFilterColumnActiveByDefaultState() {
+    const result = {};
+    if (this.table.oTableColumnsFilterComponent) {
+      result['filter-column-active-by-default'] = this.table.showFilterByColumnIcon;
     }
     return result;
   }
@@ -205,6 +217,7 @@ export class OTableStorage {
     initialConfiguration['select-column-visible'] = this.table.oTableOptions.selectColumn.visible;
     initialConfiguration['filter-case-sensitive'] = this.table.filterCaseSensitive;
     initialConfiguration['query-rows'] = this.table.originalQueryRows;
+    initialConfiguration['filter-column-active-by-default'] = this.table.originalFilterColumnActiveByDefault;
 
     result['initial-configuration'] = initialConfiguration;